+++ Neocognitron/Point2D.cs
@@ -5,20 +5,20 @@ namespace Tac.Neocognitron
 {
 	public class Point2D
 	{
-		public double X;
-		public double Y;
+		public float X;
+		public float Y;
 
-		public Point2D(double x, double y)
+		public Point2D(float x, float y)
 		{
 			X = x;
 			Y = y;
 		}
 
-		public double Distance(double x, double y)
+		public float Distance(float x, float y)
 		{
-			double dx = X - x;
-			double dy = Y - y;
-			return Math.Sqrt(dx * dx + dy * dy);
+			float dx = X - x;
+			float dy = Y - y;
+			return (float)Math.Sqrt(dx * dx + dy * dy);
 		}
 
 	}
